requirements.txt
attrs==19.1.0
autopep8==1.4.4
beautifulsoup4==4.5.3
biopython==1.73
blis==0.2.4
Brlapi==0.6.5
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cryptography==1.7.1
cssselect==1.0.1
cssutils==1.0
cupshelpers==1.0
cycler==0.10.0
cymem==2.0.2
en-core-web-sm==2.1.0
entrypoints==0.3
flake8==3.7.7
Flask==1.0.3
html5lib==0.999999999
httplib2==0.9.2
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.14.0
Jinja2==2.10.1
jsonschema==3.0.1
keyring==10.1
keyrings.alt==1.3
kiwisolver==1.1.0
louis==3.0.0
lxml==3.7.1
Mako==1.0.6
MarkupSafe==1.1.1
matplotlib==3.0.3
mccabe==0.6.1
murmurhash==1.0.2
numpy==1.16.4
pandas==0.24.2
parso==0.5.0
pexpect==4.2.1
Pillow==4.0.0
plac==0.9.6
preshed==2.0.1
ptyprocess==0.5.1
pyasn1==0.1.9
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pyflakes==2.1.1
pygobject==3.22.0
pyparsing==2.4.0
pyrsistent==0.15.1
pysmbc==1.0.15.6
python-apt==1.4.0b3
python-dateutil==2.8.0
python-debian==0.1.30
python-debianbts==2.6.1
pytz==2019.1
pyxdg==0.25
rdflib==4.2.2
regex==2017.1.17
reportbug==7.1.7
reportlab==3.3.0
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
SecretStorage==2.3.1
six==1.12.0
spacy==2.1.4
srsly==0.0.5
thinc==7.0.4
tqdm==4.31.1
unattended-upgrades==0.1
urllib3==1.24.3
virtualenv==16.6.0
wasabi==0.2.2
webencodings==0.5
Werkzeug==0.15.4
yapf==0.27.0
